added new look component
