+++ PolyLookupComponent/package.json
@@ -13,17 +13,20 @@
   },
   "dependencies": {
     "@fluentui/react": "8.29.0",
+    "@fluentui/react-components": "^9.54.13",
     "@tanstack/react-query": "^4.24.10",
     "axios": "^1.3.4",
     "handlebars": "^4.7.8",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "sprintf-js": "^1.1.3"
   },
   "devDependencies": {
     "@microsoft/eslint-plugin-power-apps": "^0.2.51",
     "@types/node": "^18.8.2",
     "@types/powerapps-component-framework": "^1.3.14",
+    "@types/react": "^16.8",
+    "@types/react-dom": "^16.8",
     "@types/sprintf-js": "^1.1.4",
     "@types/xrm": "^9.0.70",
     "@typescript-eslint/eslint-plugin": "^5.39.0",
@@ -37,7 +40,7 @@
     "pcf-scripts": "^1.33.3",
     "pcf-start": "^1.33.3",
     "prettier": "3.0.3",
-    "react": "^16.8",
+    "react": "^16.14.0",
     "typescript": "^4.8.4"
   }
 }
