--- PolyLookupComponent/selected.json
@@ -1,1 +1,0 @@
-["white", "yellow"]
