--- PolyLookupComponent/suggestions.json
@@ -1,17 +1,0 @@
-[
-  "black",
-  "blue",
-  "brown",
-  "cyan",
-  "green",
-  "magenta",
-  "mauve",
-  "orange",
-  "pink",
-  "purple",
-  "red",
-  "rose",
-  "violet",
-  "white",
-  "yellow"
-]
