+++ PolyLookupComponent/tsconfig.json
@@ -3,8 +3,6 @@
   "compilerOptions": {
     "typeRoots": ["node_modules/@types"],
     "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "module": "es2015",
-    "moduleResolution": "node"
+    "allowSyntheticDefaultImports": true
   }
 }
